Public API stabilization: rename packages and introduce internal implementations
