+++ backend/mirai-console/src/main/kotlin/net/mamoe/mirai/console/util/JavaPluginScheduler.kt
@@ -7,10 +7,14 @@
  * https://github.com/mamoe/mirai/blob/master/LICENSE
  */
 
+@file:JvmMultifileClass
+@file:JvmName("ConsoleUtils")
+
 package net.mamoe.mirai.console.util
 
-import kotlinx.coroutines.*
-import kotlinx.coroutines.future.future
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Runnable
+import net.mamoe.mirai.console.internal.util.JavaPluginSchedulerImpl
 import net.mamoe.mirai.console.plugin.jvm.JavaPlugin
 import java.util.concurrent.Callable
 import java.util.concurrent.CompletableFuture
@@ -19,16 +23,13 @@ import kotlin.coroutines.CoroutineContext
 
 
 /**
- * 拥有生命周期管理的 Java 线程池.
+ * 拥有生命周期管理的简单 Java 线程池.
  *
  * 在插件被 [卸载][JavaPlugin.onDisable] 时将会自动停止.
  *
  * @see JavaPlugin.scheduler 获取实例
  */
-public class JavaPluginScheduler internal constructor(parentCoroutineContext: CoroutineContext) : CoroutineScope {
-    public override val coroutineContext: CoroutineContext =
-        parentCoroutineContext + SupervisorJob(parentCoroutineContext[Job])
-
+public interface JavaPluginScheduler : CoroutineScope {
     /**
      * 新增一个 Repeating Task (定时任务)
      *
@@ -36,60 +37,39 @@ public class JavaPluginScheduler internal constructor(parentCoroutineContext: Co
      *
      * @see Future.cancel 取消这个任务
      */
-    public fun repeating(intervalMs: Long, runnable: Runnable): Future<Void?> {
-        return this.future {
-            while (isActive) {
-                withContext(Dispatchers.IO) { runnable.run() }
-                delay(intervalMs)
-            }
-            null
-        }
-    }
+    public fun repeating(intervalMs: Long, runnable: Runnable): Future<Void?>
 
     /**
      * 新增一个 Delayed Task (延迟任务)
      *
      * 在延迟 [delayMillis] 后执行 [runnable]
      */
-    public fun delayed(delayMillis: Long, runnable: Runnable): CompletableFuture<Void?> {
-        return future {
-            delay(delayMillis)
-            withContext(Dispatchers.IO) {
-                runnable.run()
-            }
-            null
-        }
-    }
+    public fun delayed(delayMillis: Long, runnable: Runnable): CompletableFuture<Void?>
 
     /**
      * 新增一个 Delayed Task (延迟任务)
      *
      * 在延迟 [delayMillis] 后执行 [runnable]
      */
-    public fun <R> delayed(delayMillis: Long, runnable: Callable<R>): CompletableFuture<Void?> {
-        return future {
-            delay(delayMillis)
-            withContext(Dispatchers.IO) { runnable.call() }
-            null
-        }
-    }
+    public fun <R> delayed(delayMillis: Long, runnable: Callable<R>): CompletableFuture<Void?>
 
     /**
      * 异步执行一个任务, 最终返回 [Future], 与 Java 使用方法无异, 但效率更高且可以在插件关闭时停止
      */
-    public fun <R> async(supplier: Callable<R>): Future<R> {
-        return future {
-            withContext(Dispatchers.IO) { supplier.call() }
-        }
-    }
+    public fun <R> async(supplier: Callable<R>): Future<R>
 
     /**
      * 异步执行一个任务, 没有返回
      */
-    public fun async(runnable: Runnable): Future<Void?> {
-        return future {
-            withContext(Dispatchers.IO) { runnable.run() }
-            null
-        }
+    public fun async(runnable: Runnable): Future<Void?>
+
+    public companion object {
+        /**
+         * 创建一个 [JavaPluginScheduler]
+         */
+        @JvmStatic
+        @JvmName("create")
+        public operator fun invoke(parentCoroutineContext: CoroutineContext): JavaPluginScheduler =
+            JavaPluginSchedulerImpl(parentCoroutineContext)
     }
 }
